ColorSet.cpp: relaxed the restriction to 74 colours. Action.cpp: prevented a crash with missing animations. And another batch of MSVS2005 updates.
git-svn-id: http://ctp2.darkdust.net/anonsvn/trunk@613 34d2fc0c-33df-0310-8a34-85f8040ba873

+++ ctp2_code/gfx/spritesys/FacedSprite.h
@@ -97,7 +97,7 @@ public:
 	virtual sint32	ParseFromTokens(Token *theToken);
 	virtual void	AllocateFrameArrays(size_t count);
 
-	virtual uint16	GetNumFrames(void) const    { return m_facedFrameCount; };
+	virtual size_t	GetNumFrames(void) const    { return m_facedFrameCount; };
 	virtual void	SetNumFrames(uint16 num)    { m_facedFrameCount = num; }
 
 protected:
